clicking a submission redirects to submissions page + rename a file
